// src/types/graphql.d.ts
export type GetStandingsQuery = {
  __typename?: 'query_root';
  events: Array<{
    __typename?: 'events';
    round_number?: number | null;
    name?: string | null;
  }>;
  drivers: Array<{
    __typename?: 'drivers';
    abbreviation?: string | null;
    full_name?: string | null;
    latest_constructor: Array<{
      __typename?: 'driver_sessions';
      constructor?: {
        __typename?: 'constructors';
        name?: string | null;
        color?: string | null;
      } | null;
    }>;
    driver_standings: Array<{
      __typename?: 'driver_standings';
      round?: number | null;
      points?: bigint | number | null;
      position?: number | null;
    }>;
  }>;
  constructors: Array<{
    __typename?: 'constructors';
    name?: string | null;
    color?: string | null;
    constructor_standings: Array<{
      __typename?: 'constructor_standings';
      round?: number | null;
      points?: bigint | number | null;
      position?: number | null;
    }>;
  }>;
};